Add a feature to update an information and some improvements

+++ config/admin.php
@@ -4,9 +4,14 @@
 
 	$NB_TAG_MAX = 5;
 
+	function error($msg) {
+		echo("<p class='alert-danger alert'>" . $msg . "</p>");
+	}
+
 	function displayError() {
 		global $link;
-		echo "<p class='alert-danger alert'><b>Requête SQL échouée ! : </b>" . mysqli_error($link) . "</p>";
+
+		error("<b>Requête SQL échouée ! : </b>" . mysqli_error($link));
 	}
 
 	function displaySuccess() {
@@ -27,7 +32,7 @@
 		array_unshift($keys, $idKey);
 
 		$deleteType = $table;
-		if ($_POST['delete-type'] === $deleteType) {
+		if (isset($_POST['delete-type']) && $_POST['delete-type'] === $deleteType) {
 			$deleteId = $_POST['delete-id'];
 			$sqlDelete = "DELETE FROM " . $table . " WHERE " . $idKey . " = " . $deleteId;
 			$result = mysqli_query($link, $sqlDelete);
@@ -67,39 +72,115 @@
 		echo "</tbody></table></div>";
 	}
 
-	if ($_SERVER['REQUEST_METHOD'] == 'POST'){
+	if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+		$action = $_POST['action'];
 
 		// Ajout d'une information
-		if (!empty($_POST['description']) && !empty($_POST['link']) && !empty($_POST['fieldDescription']) && !empty($_POST['categoryAuthorDescription']) && !empty($_POST['release_date'])) {
-			$missingValues = false;
+		if (!empty($_POST['description']) && !empty($_POST['link']) && !empty($_POST['fieldDescription']) && !empty($_POST['authorName']) && !empty($_POST['release_date'])) {
+			if ($action === 'insertInformation') {
+				$missingValues = false;
 
-			$sql = "INSERT INTO info.Information (description, link, field,categoryMedia,author,insert_date, release_date)	VALUES ('" . $_POST["description"] . "', '" . $_POST["link"] . "', '" . $_POST["fieldDescription"] . "', '" . $_POST["categoryMediaDescription"] . "', '" . $_POST["categoryAuthorDescription"] . "',  now(), '" . $_POST['release_date'] . "')";
-			var_dump($sql);
-			$result = mysqli_query($link, $sql);
-			$idInformation = mysqli_insert_id($link);
+				$sql = "INSERT INTO info.Information (description, link, field,categoryMedia,author,insert_date, release_date)	VALUES ('" . $_POST["description"] . "', '" . $_POST["link"] . "', '" . $_POST["fieldDescription"] . "', '" . $_POST["categoryMediaDescription"] . "', '" . $_POST["authorName"] . "',  now(), '" . $_POST['release_date'] . "')";
+				var_dump($sql);
+				$result = mysqli_query($link, $sql);
+				$idInformation = mysqli_insert_id($link);
 
-			$success = true;
-			if ($result === false) {
-				displayError();
-				$success = false;
-			} else {
-				for ($i = 1; $i <= $NB_TAG_MAX; ++$i) {
-					$tagId = $_POST['tagName' . $i];
+				$success = true;
+				if ($result === false) {
+					displayError();
+					$success = false;
+				} else {
+					for ($i = 1; $i <= $NB_TAG_MAX; ++$i) {
+						$tagId = $_POST['tagName' . $i];
 
-					if ($tagId) {
-						$sql = "INSERT INTO info.Information_tag (idInformation, idTag)	VALUES ('" . $idInformation . "', '" . $tagId . "')";
-						var_dump($sql);
-						$result = mysqli_query($link, $sql);
+						if ($tagId) {
+							$sql = "INSERT INTO info.Information_tag (idInformation, idTag)	VALUES ('" . $idInformation . "', '" . $tagId . "')";
+							var_dump($sql);
+							$result = mysqli_query($link, $sql);
 
-						if ($result === false) {
-							displayError();
-							$success = false;
+							if ($result === false) {
+								displayError();
+								$success = false;
+							}
 						}
 					}
 				}
-			}
-			if ($success){
-				displaySuccess();
+				if ($success){
+					displaySuccess();
+				}
+			} else if ($action === 'updateInformation') {
+				$missingValues = false;
+				$idInformation = $_POST["idInformation"];
+
+				$sql = "UPDATE info.Information SET description = ?, link = ?, field = ?, categoryMedia = ?, author = ?, release_date = ? WHERE idInformation = ?;";
+				var_dump($sql);
+
+				$statement = mysqli_prepare($link, $sql);
+				mysqli_stmt_bind_param($statement,
+					"ssiiisi",
+					$_POST["description"],
+					$_POST["link"],
+					$_POST["fieldDescription"],
+					$_POST["categoryMediaDescription"],
+					$_POST["authorName"],
+					$_POST["release_date"],
+					$idInformation
+				);
+				$result = mysqli_stmt_execute($statement);
+
+				mysqli_stmt_close($statement);
+
+				$success = true;
+				if ($result === false) {
+					displayError();
+					$success = false;
+				} else {
+					$sql = "DELETE FROM info.Information_tag WHERE idInformation = ?;";
+					var_dump($sql);
+
+					$statement = mysqli_prepare($link, $sql);
+					mysqli_stmt_bind_param($statement,
+						"i",
+						$idInformation
+					);
+					$result = mysqli_stmt_execute($statement);
+					mysqli_stmt_close($statement);
+
+					if ($result === false) {
+						displayError();
+						$success = false;
+					} else {
+						for ($i = 1; $i <= $NB_TAG_MAX; ++$i) {
+							$tagId = $_POST['tagName' . $i];
+
+							if ($tagId) {
+								$sql = "INSERT INTO info.Information_tag (idInformation, idTag)	VALUES ('" . $idInformation . "', '" . $tagId . "')";
+								var_dump($sql);
+								$result = mysqli_query($link, $sql);
+
+								if ($result === false) {
+									displayError();
+									$success = false;
+								}
+							}
+						}
+					}
+				}
+				if ($success){
+					displaySuccess();
+				}
+				/*
+				if ($result == false) {
+					displayError();
+				} else {
+
+
+					$_POST['tagName1']
+
+					displaySuccess();
+				}*/
+			} else {
+				error("Paramètre POST 'action' est manquant !");
 			}
 		}
 
@@ -154,7 +235,7 @@
 			}else{
 				displaySuccess();
 			}
-		}	
+		}
 
 		// Ajoût d'un nouveau tag
 		if (!empty($_POST['tag_name'])) {
@@ -169,14 +250,44 @@
 			if ($result === false) {
 				displayError();
 			}
-		}	
+		}
 
 		if ($missingValues){
 			displayWarning();
 		}
 	}
-
 ?>
+<script type="text/javascript">
+	function updateInformationSelect(idInformation){
+		const i = document.getElementById('updateInformation');
+		var j = jsArraySelectInformationUpdate[idInformation] || {};
+
+		i.querySelector('[type=submit]').disabled = !idInformation;
+
+		updateValue('description', j.description);
+		updateValue('link', j.link);
+		updateValue('fieldDescription', j.infoField);
+		updateValue('categoryMediaDescription', j.infoCategoryMedia);
+		updateValue('authorName', j.infoAuthor);
+		updateValue('release_date', j.infoReleaseDate);
+		updateValue('description', j.description);
+		updateValue('idInformation', idInformation);
+
+		for(var k = 0; k < <?php echo $NB_TAG_MAX; ?>; k++){
+			updateValue('tagName' + (k+1), "")
+		}
+
+		if (j.tags != null) {
+			j.tags.split(',').forEach(
+				(element,index) => updateValue('tagName' + (index+1), element)
+			);
+		}
+
+		function updateValue(field, value){
+			i.querySelector('[name=\'' + field + '\']').value = value || '';
+		}
+	}
+</script>
 
 <div id="content">
 	<h1>MAJ de la bdd</h1>
@@ -197,6 +308,7 @@
 
 							echo "<br><label for='fieldDescription'>Domaine :&nbsp;</label>";
 							echo "<select name='fieldDescription'>";
+							echo "<option value=''></option>";
 							while ($row = mysqli_fetch_array($result)) {
 								echo "<option value='" . $row['idField'] . "'>" . $row['description'] . "</option>";
 							}
@@ -207,6 +319,7 @@
 
 							echo "<br><label for='categoryMediaDescription'>Média :&nbsp;</label>";
 							echo "<select name='categoryMediaDescription'>";
+							echo "<option value=''></option>";
 							while ($row = mysqli_fetch_array($result)) {
 								echo "<option value='" . $row['idCategoryMedia'] . "'>" . $row['description'] . "</option>";
 							}
@@ -215,8 +328,9 @@
 							$sql = "SELECT * FROM info.Author ORDER BY name ASC;";
 							$result = mysqli_query($link, $sql);
 
-							echo "<br><label for='categoryAuthorDescription'>Nom de l'auteur :&nbsp;</label>";
-							echo "<select name='categoryAuthorDescription'>";
+							echo "<br><label for='authorName'>Nom de l'auteur :&nbsp;</label>";
+							echo "<select class='select-200' name='authorName'>";
+							echo "<option value=''></option>";
 							while ($row = mysqli_fetch_array($result)) {
 								echo "<option value='" . $row['idAuthor'] . "'>" . $row['name'] . "</option>";
 							}
@@ -240,8 +354,110 @@
 							}
 						?>
 						 <br><label for="release_date">Date de publication :&nbsp;</label>
-						 <input type="text" id="release_date" name="release_date" placeholder="yyyy-mm-dd" title="yyyy-mm-dd"size=5>
-						 <br><input type="submit" value="Envoyer">
+						 <input type="text" id="release_date" name="release_date" placeholder="yyyy-mm-dd" title="yyyy-mm-dd"size=10>
+						 <input type="hidden" name="action" value="insertInformation" />
+						 <br><input type="submit" value="Ajouter">
+					</form>
+				</p>
+			</div>			
+			<div class="col">
+				<h2>Modification d'une information</h2>
+				<p>
+					<form action="" method="post" id="updateInformation">
+						<label for="information">Information :&nbsp;</label>
+						<?php
+							$sql = "SELECT Information.*, GROUP_CONCAT(Information_tag.idTag SEPARATOR ',') as tags FROM info.Information
+								left Join Information_tag on Information_tag.idInformation = Information.idInformation
+								group by Information.idInformation
+								ORDER BY Information.idInformation ASC;";
+
+							$result = mysqli_query($link, $sql);
+							echo "<select class='select-200' onChange='updateInformationSelect(this.value)' name='Information'>";
+							echo "<option value=''></option>";
+
+							$informationArray = Array();
+							while ($row = mysqli_fetch_array($result)) {
+								echo "<option value='" . $row['idInformation'] . "'>" . $row['description'] . "</option>";
+
+								$informationArray[$row['idInformation']] = array(
+									'description' => $row['description'],
+									'link' => $row['link'],
+									'infoField' => $row['field'],
+									'infoCategoryMedia' => $row['categoryMedia'],
+									'infoAuthor' => $row['author'],
+									'infoReleaseDate' => $row['release_date'],
+									'tags' => $row['tags']
+								);
+							}
+							echo "</select>";
+
+							$jsEncodedSelectInformationUpdate = json_encode($informationArray);
+
+							echo "<script type='text/javascript'>";
+							echo "var jsArraySelectInformationUpdate = ". $jsEncodedSelectInformationUpdate . ";\n";
+							echo "</script>";
+						?>						
+
+						<label for="description">Description :&nbsp;</label>
+						<input type="text" id="description" name="description">
+						<br><label for="link">Lien :&nbsp;</label>
+						<input type="text" id="link" name="link">
+						<?php
+							$sql = "SELECT * FROM Field ORDER BY description ASC";
+							$result = mysqli_query($link, $sql);
+
+							echo "<br><label for='fieldDescription'>Domaine :&nbsp;</label>";
+							echo "<select name='fieldDescription'>";
+							echo "<option value=''></option>";
+							while ($row = mysqli_fetch_array($result)) {
+								echo "<option value='" . $row['idField'] . "'>" . $row['description'] . "</option>";
+							}
+							echo "</select>";
+
+							$sql = "SELECT * FROM CategoryMedia ORDER BY description ASC";
+							$result = mysqli_query($link, $sql);
+
+							echo "<br><label for='categoryMediaDescription'>Média :&nbsp;</label>";
+							echo "<select name='categoryMediaDescription'>";
+							echo "<option value=''></option>";
+							while ($row = mysqli_fetch_array($result)) {
+								echo "<option value='" . $row['idCategoryMedia'] . "'>" . $row['description'] . "</option>";
+							}
+							echo "</select>";
+
+							$sql = "SELECT * FROM info.Author ORDER BY name ASC;";
+							$result = mysqli_query($link, $sql);
+
+							echo "<br><label for='authorName'>Nom de l'auteur :&nbsp;</label>";
+							echo "<select class='select-200' name='authorName'>";
+							echo "<option value=''></option>";
+							while ($row = mysqli_fetch_array($result)) {
+								echo "<option value='" . $row['idAuthor'] . "'>" . $row['name'] . "</option>";
+							}
+							echo "</select>";
+
+							$sql = "SELECT * FROM info.Tag ORDER BY name ASC;";
+							$result = mysqli_query($link, $sql);
+
+							$tagArray = [];
+							while ($row = mysqli_fetch_array($result)) {
+								$tagArray[] = Array("id" => $row['idTag'], "name" => $row['name']);
+							}
+							for ($i = 1; $i <= $NB_TAG_MAX; ++$i) {
+								echo "<br><label for='tagName" . $i . "'>Tag " . $i . ": &nbsp;</label>";
+								echo "<select name='tagName" . $i . "'>";
+								echo "<option value=''></option>";
+								foreach ($tagArray as $tag) {
+									echo "<option value='" . $tag['id'] . "'>" . $tag['name'] . "</option>";
+								}
+								echo "</select>&nbsp;";
+							}
+						?>
+						 <br><label for="release_date">Date de publication :&nbsp;</label>
+						 <input type="text" id="release_date" name="release_date" placeholder="yyyy-mm-dd" title="yyyy-mm-dd"size=10>
+						 <input type="hidden" name="action" value="updateInformation" />
+						 <input type="hidden" name="idInformation" />
+						 <br><input type="submit" disabled="disabled" value="Mettre à jour">
 					</form>
 				</p>
 			</div>
@@ -257,7 +473,7 @@
 					<form action="" method="post">
 					 <label for="name">Nom :&nbsp;</label>
 					 <input type="text" id="name" name="name">
-					 <input type="submit" value="Envoyer">
+					 <input type="submit" value="Ajouter">
 					</form>
 				</p>
 			</div>
@@ -273,7 +489,7 @@
 					<form action="" method="post">
 						<label for="field_description">Nom :&nbsp;</label>
 						<input type="text" id="field_description" name="field_description">
-						<input type="submit" value="Envoyer">
+						<input type="submit" value="Ajouter">
 					</form>
 				</p>
 			</div>
@@ -289,7 +505,7 @@
 					<form action="" method="post">
 						<label for="typedemedia_description">Type de média :&nbsp;</label>
 						<input type="text" id="typedemedia_description" name="typedemedia_description">
-						<input type="submit" value="Envoyer">
+						<input type="submit" value="Ajouter">
 					</form>
 				</p>
 			</div>
@@ -306,7 +522,7 @@
 					<form action="" method="post">
 						<label for="tag_name">Nom :&nbsp;</label>
 						<input type="text" id="tag_name" name="tag_name">
-						<input type="submit" value="Envoyer">
+						<input type="submit" value="Ajouter">
 					</form>
 				</p>
 			</div>
